# geoFields/src/main/resources/static/OpenAPIspec.yaml
openapi: 3.0.3
info:
  title: GeoFields HTTP API
  description: |
    JSON-эндпоинты бэкенда GeoFields (поля, NDVI, менеджер организации).

    **Авторизация:** сессия Spring Security (форма логина). Запросы с браузера отправляйте с `credentials: include` / same-origin cookie.

    **CSRF:** для `POST` используется cookie `XSRF-TOKEN` и заголовок `X-XSRF-TOKEN` (см. Spring `CookieCsrfTokenRepository`).

    Главная `/` отдаёт HTML без данных пользователя; контекст сессии — `GET /api/session/context`.

    HTML-страницы (`/`, `/login`, `/register`, `/org/manager`, `/org/admin`, `/org/agronomist`) как разметка в спецификации не описаны.
  version: 1.4.3
servers:
  - url: http://localhost:8080
    description: Локальный сервер (см. `server.port` в application.properties)
...
paths:
...
  /get_ndvi_by_id:
    get:
      tags: [NDVI]
      summary: URL растрового слоя NDVI для одного поля
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: query
          required: true
          schema:
            type: string
          example: "1"
        - name: date
          in: query
          required: true
          description: Дата снимка (строка, как на фронте)
          schema:
            type: string
          example: "2024-06-15"
      responses:
        "200":
          description: Объект с полем url — шаблон тайлов `{z}/{x}/{y}`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdviTileUrlResponse"
        "400":
          description: Некорректный field_id или дата
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии или не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
...
  /get_all_ndvi_tile:
    get:
      tags: [NDVI]
      summary: Один слой NDVI по всем полям организации
      security:
        - sessionCookie: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
          example: "2024-06-15"
        - name: field_ids
          in: query
          required: false
          description: Id полей (повторяющийся query-параметр). Если не задан — все поля организации из БД.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: url, organizationId, date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdviOrgTileResponse"
        "400":
          description: Некорректная дата или field_ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Не авторизован
  /get_ndvi_value:
    get:
      tags: [NDVI]
      summary: Значение NDVI в точке (заглушка)
      security:
        - sessionCookie: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 55.75
        - name: lon
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 37.62
        - name: date
          in: query
          required: true
          schema:
            type: string
          example: "2024-06-15"
      responses:
        "400":
          description: Некорректная дата
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Не авторизован
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdviValueResponse"
...
  /api/org/admin/summary:
    get:
      tags: [OrgAdmin]
      summary: Список пользователей организации и CSRF для UI админа
      security:
        - sessionCookie: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrgSummary"
        "403":
          description: Нужна роль ORG_ADMIN
  /api/org/admin/role:
    post:
      tags: [OrgAdmin]
      summary: Назначить роль пользователю организации
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /api/org/admin/delete:
    post:
      tags: [OrgAdmin]
      summary: Удалить пользователя из организации
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserIdRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /api/org/admin/fields/{fieldId}:
    delete:
      tags: [OrgAdmin]
      summary: Удалить поле организации
      security:
        - sessionCookie: []
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "404":
          description: Поле не найдено в организации
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Сессия после входа через POST /login
  schemas:
...
    NdviTileUrlResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: URL шаблон с плейсхолдерами {z}/{x}/{y}
          example: "https://example.com/tiles/{z}/{x}/{y}"
    NdviOrgTileResponse:
      type: object
      required: [url, organizationName, date]
      properties:
        url:
          type: string
        organizationName:
          type: string
          description: Название организации из справочника (может быть пустой строкой)
        date:
          type: string
    NdviValueResponse:
      type: object
      properties:
        value:
          type: number
          format: double
          nullable: true
        msg:
          type: string
          nullable: true
    SimpleError:
      type: object
      properties:
        error:
          type: string
          example: "Некорректный field_id"
...
    CsrfInfo:
      type: object
      nullable: true
      properties:
        parameterName:
          type: string
          nullable: true
        headerName:
          type: string
          nullable: true
          example: "X-XSRF-TOKEN"
        token:
          type: string
          nullable: true
    ActionResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
...
    UserIdRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
          format: int64
    RoleUpdateRequest:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: integer
          format: int64
        role:
          type: string
          description: USER | AGRONOMIST | ORG_MANAGER | ORG_ADMIN
          enum: [USER, AGRONOMIST, ORG_MANAGER, ORG_ADMIN]
    OrgMemberItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        lastName:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        login:
          type: string
        email:
          type: string
        role:
          type: string
        roleLabel:
          type: string
        registrationStatus:
          type: string
        active:
          type: boolean
        currentUser:
          type: boolean
    AdminOrgSummary:
      type: object
      required: [organizationId, organizationName, currentLogin, members, csrf]
      properties:
        organizationId:
          type: integer
          format: int64
        organizationName:
          type: string
        currentLogin:
          type: string
        currentUserFullName:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/OrgMemberItem"
        csrf:
          $ref: "#/components/schemas/CsrfInfo"
...